requirements.txt
flask
flask-cors
openai
sentence-transformers
faiss-cpu
numpy
chromadb
gunicorn
python-dotenv
